+++ lib/coffee_http_client.dart
@@ -1,6 +1,8 @@
 export 'src/coffee_http_client.dart';
+export 'src/adapters/adapter.dart';
 export 'src/models/config.dart';
 export 'src/models/request.dart';
 export 'src/models/response.dart';
 export 'src/models/hooks.dart';
 export 'src/error.dart';
+export 'src/adapters/mock_adapter.dart';